test: split cases.rs into per-extension/per-license files with inline content
Previously `tests/cases.rs` referenced 16 fixture files under `tests/fixtures/`
via `include_str!`. Splitting the suite by file extension + target license
and inlining input/expected as raw string literals makes each test fully
self-contained — you can read what's being asserted without jumping to a
sibling file.

  tests/
    common/mod.rs        # HEADER_MIT, HEADER_APACHE, check_str, fix_to_string
    rust_mit.rs          # 7 tests (.rs, MIT)
    rust_apache.rs       # 3 tests (.rs, multi-line Apache-2.0)
    toml_mit.rs          # 3 tests (.toml, MIT)
    script_mit.rs        # 3 tests (cargo-script, MIT)

Rust normalises CRLF→LF inside string literals, so inlined fixtures match
LF-terminated `fix` output regardless of source-file line endings.

`cases.rs` and the entire `tests/fixtures/` tree are removed. Behaviour
covered is identical (16 tests + 1 doctest, all passing).
