--- crates/cargo_heather/tests/fixtures/rust_descriptive/expected.rs
@@ -1,4 +1,0 @@
-// Licensed under the MIT License.
-// More info about this file.
-
-fn main() {}
