--- crates/cargo_heather/tests/fixtures/rust_doc_comment/expected.rs
@@ -1,5 +1,0 @@
-// Licensed under the MIT License.
-
-//! Crate-level docs.
-
-fn main() {}
