--- crates/cargo_heather/tests/fixtures/rust_doc_comment/input.rs
@@ -1,3 +1,0 @@
-//! Crate-level docs.
-
-fn main() {}
