--- crates/cargo_heather/tests/fixtures/rust_empty/expected.rs
@@ -1,1 +1,0 @@
-// Licensed under the MIT License.
