--- crates/cargo_heather/tests/fixtures/rust_existing_comment/expected.rs
@@ -1,5 +1,0 @@
-// Licensed under the MIT License.
-
-// pre-existing comment
-
-fn main() {}
