--- crates/cargo_heather/tests/fixtures/rust_existing_comment/input.rs
@@ -1,3 +1,0 @@
-// pre-existing comment
-
-fn main() {}
