--- crates/cargo_heather/tests/fixtures/rust_mismatch/expected.rs
@@ -1,3 +1,0 @@
-// Licensed under the MIT License.
-
-fn main() {}
