--- crates/cargo_heather/tests/fixtures/rust_mismatch/input.rs
@@ -1,3 +1,0 @@
-// Licensed under the Apache License.
-
-fn main() {}
