--- crates/cargo_heather/tests/fixtures/rust_missing/input.rs
@@ -1,1 +1,0 @@
-fn main() {}
