--- crates/cargo_heather/tests/fixtures/script_mismatch/expected.rs
@@ -1,5 +1,0 @@
-#!/usr/bin/env cargo
----
-# Licensed under the MIT License.
-
-fn main() {}
