--- crates/cargo_heather/tests/fixtures/script_missing/input.rs
@@ -1,3 +1,0 @@
-#!/usr/bin/env cargo
----
-fn main() {}
