--- crates/cargo_heather/tests/fixtures/toml_mismatch/expected.toml
@@ -1,4 +1,0 @@
-# Licensed under the MIT License.
-
-[package]
-name = "foo"
