--- crates/cargo_heather/tests/fixtures/toml_missing/input.toml
@@ -1,2 +1,0 @@
-[package]
-name = "foo"
